config/config.py
import os
from dotenv import load_dotenv
from langchain_openai import ChatOpenAI

load_dotenv()

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
LLM_MODEL = ChatOpenAI(model=os.getenv("OPENAI_MODEL"), temperature=0)