requirements.txt
pandas
duckdb
pyarrow
langchain-openai
langgraph
streamlit
langchain
langchain-community
duckdb-engine
python-dotenv